Create Day 2: Operators
Given the meal price (base cost of a meal), tip percent (the percentage of the meal price being added as tip), and tax percent (the percentage of the meal price being added as tax) for a meal,
find and print the meal's total cost. 
Round the result to the nearest integer.